+++ layer-api/src/main/java/org/layer/domain/actionItem/controller/ActionItemController.java
@@ -3,7 +3,11 @@ package org.layer.domain.actionItem.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.layer.common.annotation.MemberId;
-import org.layer.domain.actionItem.api.ActionItemApi;
+import org.layer.domain.actionItem.controller.dto.CreateActionItemRequest;
+import org.layer.domain.actionItem.controller.dto.CreateActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.DeleteActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.MemberActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.SpaceActionItemResponse;
 import org.layer.domain.actionItem.dto.*;
 import org.layer.domain.actionItem.service.ActionItemService;
 import org.springframework.http.HttpStatus;
