+++ layer-api/src/main/java/org/layer/domain/actionItem/service/ActionItemService.java
@@ -2,6 +2,12 @@ package org.layer.domain.actionItem.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+
+import org.layer.domain.actionItem.controller.dto.CreateActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.DeleteActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.MemberActionItemResponse;
+import org.layer.domain.actionItem.controller.dto.SpaceActionItemElementResponse;
+import org.layer.domain.actionItem.controller.dto.SpaceActionItemResponse;
 import org.layer.domain.actionItem.dto.*;
 import org.layer.domain.actionItem.entity.ActionItem;
 import org.layer.domain.actionItem.repository.ActionItemRepository;
