--- layer-api/src/main/java/org/layer/domain/auth/exception/AuthExceptionType.java
@@ -1,29 +1,0 @@
-package org.layer.domain.auth.exception;
-
-import lombok.RequiredArgsConstructor;
-import org.layer.common.exception.ExceptionType;
-import org.springframework.http.HttpStatus;
-
-@RequiredArgsConstructor
-public enum AuthExceptionType implements ExceptionType {
-
-    /**
-     * 400
-     */
-    INVALID_SOCIAL_TYPE(HttpStatus.BAD_REQUEST, "지원하지 않는 방식의 로그인입니다."),
-    FORBIDDEN(HttpStatus.FORBIDDEN, "권한이 없습니다.");
-
-
-    private final HttpStatus status;
-    private final String message;
-
-    @Override
-    public HttpStatus httpStatus() {
-        return status;
-    }
-
-    @Override
-    public String message() {
-        return message;
-    }
-}
