+++ layer-api/src/main/java/org/layer/domain/external/controller/ExternalController.java
@@ -3,9 +3,8 @@ package org.layer.domain.external.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.layer.common.annotation.MemberId;
-import org.layer.domain.external.api.ExternalApi;
-import org.layer.domain.external.dto.ExternalRequest;
-import org.layer.domain.external.dto.ExternalResponse;
+import org.layer.domain.external.controller.dto.ExternalRequest;
+import org.layer.domain.external.controller.dto.ExternalResponse;
 import org.layer.external.ncp.dto.NcpResponse;
 import org.layer.external.ncp.service.NcpService;
 import org.springframework.http.ResponseEntity;
