+++ layer-api/src/main/java/org/layer/domain/jwt/JwtValidator.java
@@ -10,7 +10,7 @@ import org.springframework.util.StringUtils;
 
 import java.util.List;
 
-import static org.layer.domain.auth.exception.TokenExceptionType.INVALID_TOKEN;
+import static org.layer.common.exception.TokenExceptionType.INVALID_TOKEN;
 import static org.layer.domain.jwt.JwtValidationType.INVALID_JWT;
 import static org.layer.domain.jwt.JwtValidationType.VALID_JWT;
