+++ layer-api/src/main/java/org/layer/domain/jwt/service/JwtService.java
@@ -15,6 +15,9 @@ import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
 
+import static org.layer.common.exception.TokenExceptionType.*;
+import static org.layer.common.exception.TokenExceptionType.INVALID_REFRESH_TOKEN;
+import static org.layer.common.exception.TokenExceptionType.NO_REFRESH_TOKEN;
 import static org.layer.config.AuthValueConfig.ACCESS_TOKEN_EXPIRATION_TIME;
 import static org.layer.config.AuthValueConfig.REFRESH_TOKEN_EXPIRATION_TIME;
 import static org.layer.domain.auth.exception.TokenExceptionType.*;
