+++ layer-api/src/main/java/org/layer/domain/template/controller/dto/QuestionResponse.java
@@ -1,4 +1,4 @@
-package org.layer.domain.template.dto;
+package org.layer.domain.template.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
 import org.layer.domain.question.enums.QuestionType;
